refactored Code and begin with load data

+++ Day29/Password_manager/data.py
@@ -19,6 +19,7 @@ class Data:
                     self.read_file = json.load(file)
                 self.load_data_options()
                 self.load_status = True
+                # self.
                 messagebox.showinfo(title="Success!", message="File loaded successfully.")
             except FileNotFoundError:
                 messagebox.showinfo(title="Error!", message="File not found.")
@@ -27,18 +28,27 @@ class Data:
             except Exception as e:
                 messagebox.showinfo(title="Error!", message=f"An error occurred: {e}")
 
-    def save_data(self, website, user, password):
-        password = f"{password}"
-        try:
-            new_entry = {"Website": website, "Email/Username": user, "Password": password}
-            if not self.json_data_path:
-                self.open_json_file()
-            self.read_file.append(new_entry)
-            with open(self.json_data_path, "w") as read_data:
-                json.dump(self.read_file, read_data, indent=4, ensure_ascii=False)
-            messagebox.showinfo(title="Success!", message="Data saved successfully.")
-        except Exception as e:
-            messagebox.showinfo(title="Error!", message=f"An error occurred: {e}")
+    def save_data(self, website, user, password, password_encrypt):
+        password_string = f"{password_encrypt}"
+        if not all([website, user, password, password_encrypt]):
+            messagebox.showerror(title="Error!",
+                                 message="The data could not be saved. Please fill out all entries.")
+            return
+        response = messagebox.askokcancel(title="Are you sure you want to save?",
+                                          message=f"These are the details you entered: \n\nWebsite: "
+                                                  f"{website} \nEmail: {user} "
+                                                  f"\nPassword: {password} \n\nIs it ok to save?")
+        if response:
+            try:
+                new_entry = {"Website": website, "Email/Username": user, "Password": password_string}
+                if not self.json_data_path:
+                    self.open_json_file()
+                self.read_file.append(new_entry)
+                with open(self.json_data_path, "w") as read_data:
+                    json.dump(self.read_file, read_data, indent=4, ensure_ascii=False)
+                messagebox.showinfo(title="Success!", message="Data saved successfully.")
+            except Exception as e:
+                messagebox.showinfo(title="Error!", message=f"An error occurred: {e}")
 
     def load_data_options(self):
         key = "Website"
